[update]rubocopの自動修正

+++ app/controllers/public/places_controller.rb
@@ -3,19 +3,6 @@ class Public::PlacesController < ApplicationController
     @post_id = params[:post_id]
   end
 
-
-  def create
-    place_params.each do |place_param|
-      place = Place.new(place_param.merge(post_id: params[:post_id]))
-      unless place.save
-        render json: { message: "投稿に失敗しました。: #{place.errors.full_messages.join(', ')}" }, status: :unprocessable_entity
-        return
-      end
-    end
-    render json: { message: '投稿に成功しました。', redirect_url: post_path(params[:post_id]) }, status: :created
-
-  end
-
   def edit
     @place = Place.where(post_id: params[:post_id]).order(:place_num)
     @post_id = params[:post_id]
@@ -26,13 +13,21 @@ class Public::PlacesController < ApplicationController
     end
   end
 
-  def update
+  def create
+    place_params.each do |place_param|
+      place = Place.new(place_param.merge(post_id: params[:post_id]))
+      unless place.save
+        render json: { message: "投稿に失敗しました。: #{place.errors.full_messages.join(', ')}" }, status: :unprocessable_entity
+        return
+      end
+    end
+    render json: { message: '投稿に成功しました。', redirect_url: post_path(params[:post_id]) }, status: :created
+  end
 
+  def update
     client_place_ids = update_params.map { |place| place[:id].to_s.strip }.reject { |id| id == "null" || id.blank? }
     Place.where(post_id: params[:post_id]).each do |place|
-      unless client_place_ids.include?(place.id.to_s)
-        place.destroy
-      end
+      place.destroy unless client_place_ids.include?(place.id.to_s)
     end
 
     update_params.each do |place_param|
@@ -56,9 +51,7 @@ class Public::PlacesController < ApplicationController
         # 既存のレコードの更新処理
         place = Place.find_by(id: place_param[:id])
         if place
-          if params[:place][:image].present?
-            place.image.attach(params[:place][:image])
-          end
+          place.image.attach(params[:place][:image]) if params[:place][:image].present?
 
           unless place.update(
             place_name: place_param[:place_name],
@@ -79,11 +72,11 @@ class Public::PlacesController < ApplicationController
     render json: { message: '投稿を更新しました', redirect_url: post_path(params[:post_id]) }
   end
 
-
   private
+
   def place_params
-    params[:place].values.map do | place |
-      ActionController::Parameters.new(place).permit(:place_name, :comment, :image, :good, :latitude, :longitude, :place_num )
+    params[:place].values.map do |place|
+      ActionController::Parameters.new(place).permit(:place_name, :comment, :image, :good, :latitude, :longitude, :place_num)
     end
   end
